add task scene adds tasks; dropdowns don't work correctly
