+++ src/main/resources/com/tracker/task-add.fxml
@@ -13,7 +13,7 @@
 <?import javafx.scene.layout.VBox?>
 <?import javafx.scene.text.Font?>
 
-<VBox alignment="TOP_CENTER" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" spacing="12.0" xmlns="http://javafx.com/javafx/21" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.tracker.AddTaskController">
+<VBox alignment="TOP_CENTER" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" spacing="12.0" xmlns="http://javafx.com/javafx/21" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.tracker.TaskAddController">
    <children>
       <Label text="Add new task" textAlignment="CENTER">
          <font>
@@ -46,7 +46,7 @@
       <HBox alignment="CENTER" spacing="36.0">
          <children>
                   <Button fx:id="btn_cancel" mnemonicParsing="false" onAction="#cancelAction" prefHeight="26.0" prefWidth="122.0" text="Cancel" />
-                  <Button fx:id="btn_addTask" defaultButton="true" mnemonicParsing="false" onAction="#addTaskAction" prefHeight="26.0" prefWidth="122.0" text="Add Task" />
+                  <Button fx:id="btn_addTask" defaultButton="true" mnemonicParsing="false" onAction="#handleAddTask" prefHeight="26.0" prefWidth="122.0" text="Add Task" />
          </children>
       </HBox>
          </children>
